Factor: update to the official 0.99 release
Due to some mess regarding Factor nightly builds folder I am switching to the official release link. This should probably be switched back to nightly when significant changes in development are made, as I don't think anyone actually uses the official release instead of whatever the state of the development repo is. 0.99 came out literal days ago though so it is very much up to date at this point.