Apply a patch to fix support for CMake 4
The failed CI build revealed that our current checks for CMake versions prohibit builds with CMake 4, which ATO seems to be using now.  @arifthpe has opened a PR on our repository to enable this going forward in https://github.com/chapel-lang/chapel/pull/27039 but in the meantime, she has suggested this (admittedly hacky) patch to keep things working with ATO in the meantime.